:recycle: Refactored out factory methods and added openapi error messages for query param validation
