+++ fastapi_crudrouter/core/_base.py
@@ -1,10 +1,9 @@
-from typing import List, Optional, Callable, Union
+from typing import List, Optional, Callable
 
-from starlette.routing import Route
 from fastapi import APIRouter, HTTPException, Depends
 from pydantic import BaseModel, create_model
 
-models = []
+from ._utils import pagination_factory, schema_factory
 
 NOT_FOUND = HTTPException(404, 'Item not found')
 
@@ -31,9 +30,9 @@ class CRUDGenerator(APIRouter):
     ):
 
         self.schema = schema
-        self.pagination = self.pagination_factory(max_limit=paginate)
-        self.create_schema = create_schema if create_schema else self.schema_factory(self.schema, name='Create')
-        self.update_schema = update_schema if update_schema else self.schema_factory(self.schema, name="Update")
+        self.pagination = pagination_factory(max_limit=paginate)
+        self.create_schema = create_schema if create_schema else schema_factory(self.schema, name='Create')
+        self.update_schema = update_schema if update_schema else schema_factory(self.schema, name="Update")
 
         prefix = self._base_path + (self.schema.__name__.lower() if not prefix else prefix).strip('/')
         super().__init__(prefix=prefix, tags=[prefix.strip('/').capitalize()], *args, **kwargs)
@@ -108,30 +107,3 @@ class CRUDGenerator(APIRouter):
         return [
             'get_all', 'create', 'delete_all', 'get_one', 'update', 'delete_one'
         ]
-
-    @staticmethod
-    def schema_factory(schema_cls: BaseModel, pk_field_name: str = 'id', name: str = 'Create'):
-        """
-        Is used to create a CreateSchema which does not contain pk
-        """
-
-        fields = {f.name: (f.type_, ...) for f in schema_cls.__fields__.values() if f.name != pk_field_name}
-
-        name = schema_cls.__name__ + name
-        schema = create_model(name, **fields)
-        return schema
-
-    @staticmethod
-    def pagination_factory(max_limit: int = None) -> Callable:
-        def pagination(skip: int = 0, limit: int = max_limit):
-            if skip < 0:
-                raise HTTPException(422, "skip query parameter must be greater or equal to zero")
-
-            if limit is not None:
-                if limit <= 0:
-                    raise HTTPException(422, "limit query parameter must be greater then zero")
-                elif max_limit < limit:
-                    raise HTTPException(422, f"limit query parameter must be less then {max_limit}")
-
-            return {"skip": skip, "limit": limit}
-        return Depends(pagination)
