+++ fastapi_crudrouter/core/sqlalchemy.py
@@ -3,7 +3,7 @@ from typing import Callable
 from fastapi import Depends, HTTPException
 from pydantic import BaseModel
 
-from . import CRUDGenerator, NOT_FOUND, utils
+from . import CRUDGenerator, NOT_FOUND, _utils
 
 try:
     from sqlalchemy.orm import Session
@@ -25,13 +25,13 @@ class SQLAlchemyCRUDRouter(CRUDGenerator):
         self.db_model = db_model
         self.db_func = db
         self._pk: str = db_model.__table__.primary_key.columns.keys()[0]
-        self._pk_type: type = utils.get_pk_type(schema, self._pk)
+        self._pk_type: type = _utils.get_pk_type(schema, self._pk)
 
         if 'prefix' not in kwargs:
             kwargs['prefix'] = db_model.__tablename__
 
         if 'create_schema' not in kwargs:
-            kwargs['create_schema'] = self.schema_factory(schema, self._pk)
+            kwargs['create_schema'] = _utils.schema_factory(schema, self._pk)
 
         super().__init__(schema, *args, **kwargs)
